Fixed adding an artist, added adding albums
Adding an artist is pretty much done, execpt for sanitizing strings and such. Adding an album works to a very limited degree. I haven't yet added support for checking if an album artowrk artist exists/adding an album artwork artist/inserting an album that identifies an artwork artist. There are a few TODOs listed in here. Also need to sanitize all queries in adding an album.