Refactoring/additions to common_query, got start on adding a song
I refactored common_query and grouped the functions in to more related blocks so the could be found more easily. Also added to this file all of the functions required to insert the song, and create all relationships between songs and other entities. Have not created the function to get a song's ID.
Got (a very small) start on adding a song to the database